Relocate utils/weekday.py to standard provider
The weekday util has been created and used exclusively by the weekday operator/sensors that were already relocated to standard provider.
