+++ airflow/example_dags/example_bash_decorator.py
@@ -22,9 +22,9 @@ import pendulum
 from airflow.decorators import dag, task
 from airflow.exceptions import AirflowSkipException
 from airflow.providers.standard.operators.empty import EmptyOperator
+from airflow.providers.standard.utils.weekday import WeekDay
 from airflow.sdk import chain
 from airflow.utils.trigger_rule import TriggerRule
-from airflow.utils.weekday import WeekDay
 
 
 @dag(schedule=None, start_date=pendulum.datetime(2023, 1, 1, tz="UTC"), catchup=False)
