+++ providers/standard/src/airflow/providers/standard/sensors/weekday.py
@@ -20,9 +20,9 @@ from __future__ import annotations
 from collections.abc import Iterable
 from typing import TYPE_CHECKING
 
+from airflow.providers.standard.utils.weekday import WeekDay
 from airflow.sensors.base import BaseSensorOperator
 from airflow.utils import timezone
-from airflow.utils.weekday import WeekDay
 
 if TYPE_CHECKING:
     try:
@@ -54,7 +54,7 @@ class DayOfWeekSensor(BaseSensorOperator):
     **Example** (with :class:`~airflow.utils.weekday.WeekDay` enum): ::
 
         # import WeekDay Enum
-        from airflow.utils.weekday import WeekDay
+        from airflow.providers.standard.utils.weekday import WeekDay
 
         weekend_check = DayOfWeekSensor(
             task_id="weekend_check",
